Miscellaneous
Summary:
__str__ of module

add info of container cases in workboard shop module

fix l10n and form error in container places management

Test Plan: .

Reviewers: pastys

Differential Revision: http://phabricator.iresam.org/D829

+++ shops/forms.py
@@ -188,7 +188,7 @@ class ShopContainerCaseForm(forms.Form):
         widget=forms.HiddenInput(),
         required=False
     )
-    percentage = forms.IntegerField(
+    percentage = forms.FloatField(
         label='percentage',
         widget=forms.HiddenInput(),
         required=False
